+++ src/test/java/stringcalculator/NumbersTest.java
@@ -2,30 +2,17 @@ package stringcalculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.*;
 
 class NumbersTest {
 
-    @ParameterizedTest
-    @DisplayName(value = "구분자 , 혹은 : 문자열을 전달하면 int[] 로 반환된다.")
-    @ValueSource(strings = {"1,2:3", "1,2,3", "1:2:3"})
-    void defaultDelimiterSplit(String text) {
+    @DisplayName(value = "String[]을 int[]로 반환된다.")
+    @Test
+    void defaultDelimiterSplit() {
         int[] check = {1,2,3};
-        int[] result = Numbers.textToNumbers(text);
-
-        assertArrayEquals(result, check);
-    }
-
-    @DisplayName(value = "//와 \\n 문자 사이에 커스텀 구분자 문자열을 전달하면 int[] 로 반환된다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"//;\n1;2;3", "//!\n1!2!3", "//@\n1@2@3"})
-    void customDelimiterSplit(String text) {
-        int[] check = {1,2,3};
-        int[] result = Numbers.textToNumbers(text);
+        int[] result = Numbers.toNumbers(new String[]{"1", "2", "3"});
 
         assertArrayEquals(result, check);
     }
@@ -34,6 +21,6 @@ class NumbersTest {
     @Test
     void negative() {
         assertThatExceptionOfType(RuntimeException.class)
-                .isThrownBy(() -> Numbers.textToNumbers("-1"));
+                .isThrownBy(() -> Numbers.toNumbers(new String[]{"-1"}));
     }
 }
